change from station to stationorder in order table: 2 related function points: create order cancel order

+++ src/main/java/com/company/project/dao/TimeDAO.java
@@ -12,9 +12,9 @@ import java.util.List;
 public interface TimeDAO extends Mapper<Time> {
 
     List<Time> findTrain(Date date, String start, String end);
-    BigDecimal findStationOrder(String trainnumber, LocalDate traindate, String station);
-    void sell(String trainnumber, LocalDate traindate, BigDecimal startOrder, BigDecimal arriveOrder, int num); //卖出订单
-    void back(String trainnumber, LocalDate traindate, BigDecimal startOrder, BigDecimal arriveOrder, int num); //取消订单
-    BigDecimal findRemainSeat(String trainnumber, LocalDate traindate, BigDecimal stationorder);
+    int findStationOrder(String trainnumber, LocalDate traindate, String station);
+    void sell(String trainnumber, LocalDate traindate, int startOrder, int arriveOrder, int num); //卖出订单
+    void back(String trainnumber, LocalDate traindate, int startOrder, int arriveOrder, int num); //取消订单
+    BigDecimal findRemainSeat(String trainnumber, LocalDate traindate, int stationorder);
 
 }
