+++ src/main/java/com/company/project/service/TimeService.java
@@ -15,6 +15,6 @@ import java.util.List;
 public interface TimeService extends Service<Time> {
 
     List<Time> searchTrain(Date date, String start, String arrive);
-    boolean isTicketLeft(String trainnumber, LocalDate traindate, BigDecimal stationorder);
+    boolean isTicketLeft(String trainnumber, LocalDate traindate, int stationorder);
 
 }
